+++ src/formatter/mod.rs
@@ -98,7 +98,8 @@
 //!
 //! let highlight_lines = HighlightLines {
 //!     lines: vec![1..=1, 3..=4],  // Highlight lines 1, 3, and 4
-//!     style: HighlightLinesStyle::Theme,  // Use theme's cursorline style
+//!     style: Some(HighlightLinesStyle::Theme),  // Use theme's cursorline style
+//!     class: None,
 //! };
 //!
 //! let formatter = HtmlInlineBuilder::new()