+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Added
+- Add `class` field to `HighlightLines` in HTML inline formatter for custom CSS classes on highlighted lines
+- Append CSS style rules `transition: background-color .5s`, `width: 100%`, `display: inline-block` to each `<span>` line on `HtmlInline` formatter when `HighlightStyle:Theme`
+
+### Changed
+- **Breaking** Change `HighlightLines.style` field from `HighlightLinesStyle` to `Option<HighlightLinesStyle>` allowing users to define either style or class for highlighted lines
+
 ## [0.5.0] - 2025-08-07
 
 ### Changed